Add a CustomString type to allow unescaped output.
When trying to print multi-line (or otherwise formatted) messages, the
fact that all strings are escaped becomes a problem. For example:

```
func PanicLog() {
    if err := recover(); err != nil {
        log.Crit("Panic", "stack", string(debug.Stack()))
    }
}
```
The stack trace becomes a single line, with all \n's and \t's escaped.
With this new type, the central line becomes:
```
log.Crit("Panic", "stack", log.CustomString(debug.Stack()))
```
and now it formats in a readable way.

CustomString multi-line values are rendered with a leading `> `, to
enable machine parsing. An example output is viewable here:

https://gist.github.com/kormat/10e09e4dcfef0115c65fcdc20c98297e